refactor(hub): rename HubView.tsx to hub-view.tsx for kebab-case file names
