+++ src/renderer/components/Hub/hub-view.tsx
